third_party/WebKit/LayoutTests/http/tests/loading/pending-script-leak.html

<html>
<body>
<script>
var iframe = document.createElement("iframe");
document.body.appendChild(iframe);

var s = iframe.contentWindow.document.createElement("script");
s.src = "http://127.0.0.1:8000/resources/slow-script.pl?delay=10000";
s.async = false;
iframe.contentWindow.document.body.appendChild(s);

if (window.testRunner)
    testRunner.dumpAsText();
</script>
<p>There should be no leaks when run with --enable-leak-detection
</body>
</html>
